added file structure, added .idea to gitignore
